+++ savvy-ffi/src/lib.rs
@@ -10,6 +10,9 @@ pub const Rboolean_FALSE: Rboolean = 0;
 pub const Rboolean_TRUE: Rboolean = 1;
 pub type Rboolean = ::std::os::raw::c_int;
 
+#[cfg(feature = "complex")]
+pub use num_complex::Complex64;
+
 // SEXP
 pub type SEXP = *mut ::std::os::raw::c_void;