+++ src/sexp/real.rs
@@ -200,6 +200,12 @@ impl OwnedRealSexp {
         out.as_mut_slice().copy_from_slice(x_slice);
         Ok(out)
     }
+
+    /// Constructs a new integer vector from a scalar value.
+    pub fn try_from_scalar(value: f64) -> crate::error::Result<Self> {
+        let sexp = unsafe { crate::unwind_protect(|| savvy_ffi::Rf_ScalarReal(value))? };
+        Self::new_from_raw_sexp(sexp, 1)
+    }
 }
 
 impl Drop for OwnedRealSexp {
@@ -253,8 +259,7 @@ impl TryFrom<f64> for OwnedRealSexp {
     type Error = crate::error::Error;
 
     fn try_from(value: f64) -> crate::error::Result<Self> {
-        let sexp = unsafe { crate::unwind_protect(|| savvy_ffi::Rf_ScalarReal(value))? };
-        Self::new_from_raw_sexp(sexp, 1)
+        Self::try_from_scalar(value)
     }
 }